+++ WindowsPcl/AdjustEventFailure.cs
@@ -9,15 +9,17 @@ namespace AdjustSdk
         public string Adid { get; set; }
         public string EventToken { get; set; }
         public bool WillRetry { get; set; }
+        public string CallbackId { get; set; }
         public Dictionary<string, string> JsonResponse { get; set; }
 
         public override string ToString()
         {
-            return Pcl.Util.F("Event Failure msg:{0} time:{1} adid:{2} event:{3} retry:{4} json:{5}",
+            return Pcl.Util.F("Event Failure msg:{0} time:{1} adid:{2} event:{3} cid:{4} retry:{5} json:{6}",
                 Message,
                 Timestamp,
                 Adid,
                 EventToken,
+                CallbackId,
                 WillRetry,
                 JsonResponse);
         }